support for 3.16
There is a problem with the extension though. Everything works as expected and you can remove all the sections that you want to with no problem, however for some reason after you reboot your computer it displays an "Error" in the Tweak Tool listing. Everything continues to work just fine though so I have no idea what the error is about.